requirements.txt
Flask
torch
Pillow
werkzeug
